password fallback for legacy images
