+++ tests/Testcontainers.OpenSearch.Tests/Usings.cs
@@ -1,4 +1,5 @@
 global using System;
+global using System.Linq;
 global using System.Threading.Tasks;
 global using DotNet.Testcontainers.Commons;
 global using JetBrains.Annotations;
